feat: renamed documents ui folder
