+++ components/documents/FilePreview.tsx
